delegated methods from reservation_manager to lowerlevel classes. changed some naming conventions and moved tests to their new test location.
