+++ lib/reservation_dates.rb
@@ -1,7 +1,7 @@
 require 'date'
 
 module HotelBookings
-  class Reservation_Dates
+  class ReservationDates
     attr_reader :checkin, :checkout
 
     def initialize(checkin:, checkout:)
@@ -12,6 +12,15 @@ module HotelBookings
     def total_nights
       number_of_nights = @checkout - @checkin 
       return number_of_nights.to_f
+    end 
+
+    def reservation_nights
+      nights = []
+      total_nights.each do |night| 
+        nights.push(night)
+        night += 1
+      end 
+      return nights 
     end 
   end 
 end
